stop deleting older queries if they are persistent

+++ www/cgi-bin/interrogar.py
@@ -44,7 +44,7 @@ def sendRequestInterrogar():
 			else:
 				date = datetime.fromisoformat(query_records[filename_json_id]['datetime'])
 				time_passed = datetime.now() - date
-				if time_passed.days >= 2:
+				if time_passed.days >= 2 and query_records[filename_json_id]['persistent'] == False:
 					os.remove(os.path.join(json_path, filename))
 	
 	conllu_json = []
@@ -88,8 +88,7 @@ def sendPOSTInterrogar():
 	criterio, parametros, conllu, nomePesquisa, script = definirVariaveisDePesquisa(cgi.FieldStorage())
 
 	caminhoCompletoConllu = "./interrogar-ud/conllu/" + conllu
-	dataAgora = str(datetime.now()).replace(' ', '_').split('.')[0]
-	caminhoCompletoHtml = './interrogar-ud/resultados/' + slugify(nomePesquisa) + '_' + dataAgora.replace(':', '_') + '.html'
+	dataAgora = str(datetime.now()).replace(' ', '_')
 
 	if nomePesquisa and nomePesquisa not in fastSearch:
 		try:
@@ -103,10 +102,7 @@ def sendPOSTInterrogar():
 
 	numeroOcorrencias, casosOcorrencias, fullParameters, json_id, scriptParams = realizarBusca(conllu, caminhoCompletoConllu, int(criterio), parametros, script, fast)
 	
-	#parametros_antigos = None
-	#if script:
-		#parametros_antigos = parametros
-		#parametros = fullParameters
+	caminhoCompletoHtml = './interrogar-ud/resultados/' + slugify(nomePesquisa) + '-' + json_id + '.html'
 
 	arquivoHtml = paginaHtml(caminhoCompletoConllu, caminhoCompletoHtml, nomePesquisa, dataAgora, conllu, criterio, parametros if not script else scriptParams, numeroOcorrencias, casosOcorrencias, script, fullParameters, json_id).montarHtml()
 
@@ -234,7 +230,7 @@ class paginaHtml():
 
 		refazerPesquisa = '<br><span class="translateHtml">Refazer busca</span></a>'
 		arquivoHtml = arquivoHtml.replace('<p>critério y#z#k&nbsp;&nbsp;&nbsp; arquivo_UD&nbsp;&nbsp;&nbsp; <span id="data">data</span>&nbsp;&nbsp;&nbsp;', '<div><div style="margin-top:2px; margin-bottom:2px;"><span class="translateHtml" style="font-weight:bold">Busca:</span> <span id=expressao style="word-break: break-all">' + web.escape(self.parametros) + '</span>' + f'</div><span style="font-weight:bold">Corpus:</span> <span id=corpus>' + self.conllu + '</span><br><br>[<a href="#" class="newQueryFromResults translateHtml">Tentar outra busca</a>]' + (' [<a href="#" onclick="document.location.href = $(\'.refazerPesquisa\').attr(\'href\') + \'&save=True\';" class="translateHtml">Salvar busca</a>]' if self.nomePesquisa in fastSearch else "") + ' [<a class="refazerPesquisa translateHtml" href="../../cgi-bin/interrogar.py?corpus=' + self.conllu + '&params=' + encodeUrl(self.parametros) + '">Voltar</a>]<br><br><span id=data><small>' + prettyDate(self.dataAgora.replace('_', ' ')).beautifyDateDMAH() + '</small></span></div>')
-		arquivoHtml = arquivoHtml.replace('id="apagar_link" value="link1"', 'id=apagar_link value="' + slugify(self.nomePesquisa) + '_' + self.dataAgora.replace(':', '_') + '"')
+		arquivoHtml = arquivoHtml.replace('id="apagar_link" value="link1"', 'id=apagar_link value="' + slugify(self.nomePesquisa) + '-' + self.json_id + '"')
 		arquivoHtml = arquivoHtml.replace('<input type="hidden" id="jsonId" value="0">', '<input type="hidden" id="jsonId" name="jsonId" value="%s">' % self.json_id)
 
 		return arquivoHtml
@@ -265,8 +261,8 @@ class paginaHtml():
 		with open("./interrogar-ud/resultados/link1.html", "r") as f:
 			self.arquivoHtml = f.read()
 			self.arquivoHtml = self.arquivoHtml.replace("../cgi-bin/modelo_script.py?crit=&params=", f"../cgi-bin/modelo_script.py?crit={self.criterio}&params={encodeUrl(self.fullParameters)}")
-			self.arquivoHtml = self.arquivoHtml.replace('../cgi-bin/filtrar.py', '../cgi-bin/filtrar.py?html=' + slugify(self.nomePesquisa) + '_' + self.dataAgora.replace(':', '_') + '&udoriginal=' + self.conllu)
-			self.arquivoHtml = self.arquivoHtml.replace('../cgi-bin/conllu.py', '../cgi-bin/conllu.py?html=./interrogar-ud/resultados/' + slugify(self.nomePesquisa) + '_' + self.dataAgora.replace(':', '_') + '.html')
+			self.arquivoHtml = self.arquivoHtml.replace('../cgi-bin/filtrar.py', '../cgi-bin/filtrar.py?html=' + slugify(self.nomePesquisa) + '_' + self.json_id + '&udoriginal=' + self.conllu)
+			self.arquivoHtml = self.arquivoHtml.replace('../cgi-bin/conllu.py', '../cgi-bin/conllu.py?html=./interrogar-ud/resultados/' + slugify(self.nomePesquisa) + '_' + self.json_id + '.html')
 
 		self.arquivoHtml = self.adicionarHeader()
 		#if not self.script:
